requirements.txt
# Core Web Framework
fastapi==0.109.0
uvicorn==0.27.0
jinja2==3.1.2
python-multipart==0.0.6
aiofiles==23.2.1

# Azure Storage
azure-storage-blob==12.19.0
azure-identity==1.15.0
azure-keyvault-secrets==4.7.0
azure-servicebus==7.11.4

# Database
sqlalchemy==2.0.25
alembic==1.13.1

# Configuration and Environment
python-dotenv==1.0.0
pydantic-settings==2.1.0

# HTTP Requests
requests==2.31.0
httpx==0.26.0

# Email Processing
email-validator==2.1.0

# OCR and Document Processing
Pillow==10.2.0
PyPDF2==3.0.1
pdfplumber==0.10.3
pytesseract==0.3.10
PyMuPDF==1.23.18
PyPDF2==3.0.1
pdf2image==1.17.0
pytesseract==0.3.10

# AI/ML
openai==1.10.0

# Background Processing
celery==5.3.6
redis==5.0.1

# Data Processing
databricks-sql-connector==3.0.0

# Azure Functions (if needed)
azure-functions==1.18.0

# Testing
pytest==8.0.0
pytest-asyncio==0.23.4

# Development Tools
black==24.1.1
flake8==7.0.0
mypy==1.8.0
